small changes from email for green card and O-1
ii.MEMBERSHIP

Documentation of the alien's membership in associations in the field
for which classification is sought, which requires outstanding
achievements of their members, as judged by recognized national or
international experts in their disciplines or fields; the trend here is
a request for proof that the membership is very selective.
“The" should be capitalized.

2.

vi. SCHOLARLY ARTICLES
Evidence of the alien's authorship of scholarly articles in the field,
in professional or major trade publications or other major media;
Scholarly articles are the type of articles that a professor publishes
in a scientific journal. Articles that an art critic writes in a
magazine do not satisfy this category. Similarly catalog essays do not
satisfy this category. The questions to ask include: is the article
published in a scholarly publication? Does the article have footnotes?
USCIS has a strong “nose” for what is and is not a scholarly article.
They have many applications by scientists and professors. They know
what a scholarly article is. They will not give the benefit of the
doubt if this article is not a legitimate scholarly article.
“Is” should be capitalized

3.
vii. EXHIBITIONS AND SHOWCASES
Evidence of the display of the alien’s work in the field at artistic
exhibitions or showcases; TThe adjudicative trend is that artists are
generally granted this category. There has to be more than one
exhibition to qualify. There should be substantial evidence to prove
that the alien exhibited their work and that it was an artistic
exhibition. This category generally does not work as well for musicians
and fashion designers. USCIS explains that concerts or fashion shows
are what professionals in these fields do as part of their day-to-day
activity. That being said, a professional artist with many shows would
most likely qualify for this category, especially if there are a number
of solo exhibitions. Evidence of museum shows, biennials or high-end
gallery solo shows will also help with the second part of the test.
“The”

4.
viii. CRITICAL ROLE
Evidence that the alien has performed in a leading or critical role for
organizations or establishments that have a distinguished reputation;
some occupations are more suited to this category then others. Any
occupation where the alien has worked for an employer for many years
would have a stronger chance of satisfying this category. For example,
an architect or graphic designer, or a curator with a museum might be
able to document a strong critical role with an organization. We have
had success with artists who have gallery representation if we can
prove the artist has had a leading role with that gallery by having
multiple high profile exhibitions. One time exhibitions or one-time
involvements will generally not succeed with this category these days.
“Some”

5.

ix. HIGH REMUNERATION
Evidence that the alien has commanded a high salary or other
significantly high remuneration for services, in relation to others in
the field; in some occupations, for example Art Directors in Ad Firms,
the alien receives a salary. There are online surveys, which can
document whether the salary is high compared to others. Artists can
also get letters from their galleries if they earn a large amount from
that gallery over time. One time payments or sales would most likely
not satisfy this category.
“In”  cap

, for example,

(comma after example)
7.  To satisfy the Final Merits Determination the alien should put
their best application together so that immigration can review all of
the evidence and approve the case. This part is relatively subjective
and USCIS has great discretion in determining whether an alien is
successful. The stronger the evidence is in terms of an alien’s
achievements in the field compared to others, the more efficiently and
accurately USCIS can process the petition.
Ultimately it is this risk of failing the Final Merits Determination
that can foil what would otherwise be a successful application.

Can you move the last sentence so it is part of the paragraph?  There
is a space after "efficiently and accurately USCIS can process the
petition. “

8.

To some extent a letter or memo from the attorney can help explain why
the alien qualifies to be approved. Realistically, however, the alien
should understand that there is risk in this type of application and
that even if they feel they clearly qualify in the I-140 part A they
nevertheless could be unsuccessful in Part B.

To some extent a letter or memo from the attorney can help explain why
the alien qualifies to be approved. Realistically, however, the alien
should understand that there is risk in this type of application and
that even if they feel they clearly qualify in the I-140  Subpart A
they nevertheless could be unsuccessful in Subpart B.
(Change part A to Subpart A and part b to Subpart B)

9.
DETERMINATION

Sending the file to USCIS
When USCIS receives the file they will either approve the application,
deny the application or ask for more information so that they can
either approve or deny the application. The request for more
information is called Request for Evidence. If the application is
approved the applicant can continue to the second part of the
application.

“is called a Request for Evidence”  (a added before Request)
x. COMMMERCIAL SUCCESS

Please also remove these two sentences:
If the above standards do not readily apply to the beneficiary's
occupation, the petitioner may submit comparable evidence to establish
the beneficiary's eligibility. The adjudicative trend is that
immigration will not give weight to this category unless the occupation
does not suit itself to the above ten categories.

Technically, a semicolon does not call for capitalization, but the
convention established here does. These two details could get
capitalization to match the other eight.

viii. CRITICAL ROLE
Evidence that the alien has performed in a leading or critical role for
organizations or establishments that have a distinguished reputation;
some occupations ...

Some gets capitalized.

ix. HIGH REMUNERATION
Evidence that the alien has commanded a high salary or other
significantly high remuneration for services, in relation to others in
the field; in some occupations,

In gets capitalized.