sections 6 and 9 typos under greencard
6.
x. COMMMERCIAL SUCCESS
Evidence of commercial successes in the performing arts, as shown by
box office receipts or record, cassette, compact disk, or video sales.
This category applies to musicians, movie stars, film directors and
others whose work might appear oncharts (Billboard 100 for example).

x. COMMMERCIAL SUCCESS
Evidence of commercial successes in the performing arts, as shown by
box office receipts or record, cassette, compact disk, or video sales.
This category applies to musicians, movie stars, film directors and
others whose work might appear on charts. (Billboard 100 would be an
example).

9.
DETERMINATION

Sending the file to USCIS
When USCIS receives the file they will either approve the application,
deny the application or ask for more information so that they can
either approve or deny the application. The request for more
information is called Request for Evidence. If the application is
approved the applicant can continue to the second part of the
application.

“is called a Request for Evidence”  (a added before Request)